Verified by Marti4n v1.2.6

+++ controllers/admin.js
@@ -1,4 +1,5 @@
-const { validationResult } = require('express-validator/check');
+const fileHelper = require("../util/file");
+const { validationResult } = require('express-validator');
 const Product = require('../models/product');
 
 exports.getAddProduct = (req, res, next) => {
@@ -14,9 +15,24 @@ exports.getAddProduct = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
+  if(!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/edit-product',
+      editing: false,
+      hasError:true,
+      product: {
+        title:title,
+        price:price,
+        description:description,
+      },
+      errorMessage: "Attached file is not an image!",
+      validationErrors: []
+    });
+  }
   const errors = validationResult(req);
   if(!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -26,7 +42,6 @@ exports.postAddProduct = (req, res, next) => {
       hasError:true,
       product: {
         title:title,
-        imageUrl:imageUrl,
         price:price,
         description:description,
       },
@@ -38,8 +53,8 @@ exports.postAddProduct = (req, res, next) => {
   const product = new Product({
     title: title,
     price: price,
-    description: description,
     imageUrl: imageUrl,
+    description: description,
     userId: req.user
   });
   product
@@ -81,7 +96,7 @@ exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const updatedTitle = req.body.title;
   const updatedPrice = req.body.price;
-  const updatedImageUrl = req.body.imageUrl;
+  const image = req.file;
   const updatedDesc = req.body.description;
   const errors = validationResult(req);
   if(!errors.isEmpty()) {
@@ -92,7 +107,6 @@ exports.postEditProduct = (req, res, next) => {
       hasError:true,
       product: {
         title:updatedTitle,
-        imageUrl:updatedImageUrl,
         price:updatedPrice,
         description:updatedDesc,
         _id:prodId
@@ -111,12 +125,14 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       product.description = updatedDesc;
-      product.imageUrl = updatedImageUrl;
-      return product.save()
-      .then(result => {
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
+      return product.save().then(result => {
         console.log('UPDATED PRODUCT!');
         res.redirect('/admin/products');
-      })
+      });
     })
     .catch(err => console.log(err));
 };
@@ -136,10 +152,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
+  Product.findById(prodId)
+  .then(product => {
+    if (!product) {
+      return next(new Error('Product not found.'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id });
+  })
+  .then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
     .catch(err => console.log(err));
 };
